+++ test/foundry/utils/ProtocolHelpers.sol
@@ -20,7 +20,7 @@ contract ProtocolHelpers is TestHelpers, TestParameters {
     function _createSingleItemMakerAskOrder(
         uint112 askNonce,
         uint112 subsetNonce,
-        uint16 strategyId,
+        uint256 strategyId,
         uint8 assetType,
         uint256 orderNonce,
         address collection,
@@ -56,7 +56,7 @@ contract ProtocolHelpers is TestHelpers, TestParameters {
     function _createMultiItemMakerAskOrder(
         uint112 askNonce,
         uint112 subsetNonce,
-        uint16 strategyId,
+        uint256 strategyId,
         uint8 assetType,
         uint256 orderNonce,
         address collection,
@@ -88,7 +88,7 @@ contract ProtocolHelpers is TestHelpers, TestParameters {
     function _createSingleItemMakerBidOrder(
         uint112 bidNonce,
         uint112 subsetNonce,
-        uint16 strategyId,
+        uint256 strategyId,
         uint8 assetType,
         uint256 orderNonce,
         address collection,
@@ -123,7 +123,7 @@ contract ProtocolHelpers is TestHelpers, TestParameters {
     function _createMultiItemMakerBidOrder(
         uint112 bidNonce,
         uint112 subsetNonce,
-        uint16 strategyId,
+        uint256 strategyId,
         uint8 assetType,
         uint256 orderNonce,
         address collection,
